Implement new slicezone
