+++ .slicemachine/assets/slices/EmailSignup/mocks.json
@@ -1,21 +1,42 @@
 [
   {
-    "variation": "default-slice",
-    "name": "Default slice",
+    "variation": "DefaultSlice",
+    "name": "DefaultSlice",
     "slice_type": "email_signup",
     "items": [],
     "primary": {
-      "title": [
+      "sectionTitle": [
         {
           "type": "heading1",
-          "text": "Engage open-source metrics",
+          "text": "Utilize mission-critical architectures",
           "spans": []
         }
       ],
       "description": [
         {
           "type": "paragraph",
-          "text": "Ipsum velit excepteur ullamco ex ut culpa aliquip qui ex laborum.",
+          "text": "Consectetur qui consectetur proident sit eu eiusmod. Mollit minim pariatur esse id enim nisi ad exercitation proident consequat dolor cillum incididunt commodo. Culpa enim ut eu qui ea ex nostrud laborum.",
+          "spans": []
+        }
+      ],
+      "inputLabel": [
+        {
+          "type": "paragraph",
+          "text": "Quis veniam duis ut sint incididunt anim veniam.",
+          "spans": []
+        }
+      ],
+      "inputPlaceholder": [
+        {
+          "type": "paragraph",
+          "text": "Velit et exercitation elit adipisicing est enim eu duis ex. Eu enim nisi adipisicing in Lorem incididunt ut aliqua non aliquip. Culpa aliquip occaecat pariatur cillum incididunt.",
+          "spans": []
+        }
+      ],
+      "buttonText": [
+        {
+          "type": "paragraph",
+          "text": "Est consequat do commodo exercitation pariatur deserunt laboris aliqua.",
           "spans": []
         }
       ]
