--- .slicemachine/assets/slices/TextInfoTwo/index.stories.js
@@ -1,9 +1,0 @@
-import MyComponent from '../../../../slices/TextInfoTwo';
-
-export default {
-  title: 'slices/TextInfoTwo'
-}
-
-
-export const _DefaultSlice = () => <MyComponent slice={{"variation":"default-slice","name":"Default slice","slice_type":"text_info_two","items":[],"primary":{"sectionTitle":[{"type":"heading1","text":"Whiteboard dynamic bandwidth","spans":[]}],"leftColumnText":[{"type":"paragraph","text":"Do anim mollit occaecat nisi sunt consequat sit laborum esse ea. Mollit quis commodo nisi consectetur veniam.","spans":[]}],"rightColumnText":[{"type":"paragraph","text":"Duis aliquip amet do eu elit veniam qui. Eiusmod elit commodo minim Lorem ipsum nulla duis officia amet sunt cillum in.","spans":[]}]},"id":"_DefaultSlice"}} />
-_DefaultSlice.storyName = 'Default slice'
