+++ .slicemachine/libraries-state.json
@@ -37,8 +37,10 @@
                 "inputLabel": {
                   "type": "StructuredText",
                   "config": {
-                    "single": "paragraph",
-                    "label": "input_label"
+                    "label": "Input Label",
+                    "placeholder": "",
+                    "allowTargetBlank": true,
+                    "single": "paragraph"
                   }
                 },
                 "inputPlaceholder": {
@@ -55,28 +57,50 @@
                     "label": "button_text"
                   }
                 }
-              }
+              },
+              "items": {}
             }
           ]
         },
         "mocks": {
-          "default-slice": {
-            "variation": "default-slice",
-            "name": "Default slice",
+          "DefaultSlice": {
+            "variation": "DefaultSlice",
+            "name": "DefaultSlice",
             "slice_type": "email_signup",
             "items": [],
             "primary": {
-              "title": [
+              "sectionTitle": [
                 {
                   "type": "heading1",
-                  "text": "Engage open-source metrics",
+                  "text": "Utilize mission-critical architectures",
                   "spans": []
                 }
               ],
               "description": [
                 {
                   "type": "paragraph",
-                  "text": "Ipsum velit excepteur ullamco ex ut culpa aliquip qui ex laborum.",
+                  "text": "Consectetur qui consectetur proident sit eu eiusmod. Mollit minim pariatur esse id enim nisi ad exercitation proident consequat dolor cillum incididunt commodo. Culpa enim ut eu qui ea ex nostrud laborum.",
+                  "spans": []
+                }
+              ],
+              "inputLabel": [
+                {
+                  "type": "paragraph",
+                  "text": "Quis veniam duis ut sint incididunt anim veniam.",
+                  "spans": []
+                }
+              ],
+              "inputPlaceholder": [
+                {
+                  "type": "paragraph",
+                  "text": "Velit et exercitation elit adipisicing est enim eu duis ex. Eu enim nisi adipisicing in Lorem incididunt ut aliqua non aliquip. Culpa aliquip occaecat pariatur cillum incididunt.",
+                  "spans": []
+                }
+              ],
+              "buttonText": [
+                {
+                  "type": "paragraph",
+                  "text": "Est consequat do commodo exercitation pariatur deserunt laboris aliqua.",
                   "spans": []
                 }
               ]
@@ -88,7 +112,13 @@
           "isDirectory": true,
           "extension": "js"
         },
-        "screenshotPaths": {}
+        "screenshotPaths": {
+          "DefaultSlice": {
+            "path": "/Users/samlittlefair/Documents/boilerplates-examples/nextjs-multi-language-site/.slicemachine/assets/slices/EmailSignup/DefaultSlice/preview.png",
+            "width": 800,
+            "height": 308
+          }
+        }
       },
       "FullWidthImage": {
         "library": "slices",
