+++ components/index.js
@@ -3,4 +3,3 @@ export { default as Navigation } from './Navigation'
 export { default as LanguageSwitcher } from './LanguageSwitcher'
 export { default as Footer } from './Footer'
 export { default as Layout } from './Layout'
-export { default as SliceZone } from './SliceZone'
