--- components/slices/index.js
@@ -1,5 +1,0 @@
-export { default as EmailSignup } from './EmailSignup'
-export { default as FullWidthImage } from './FullWidthImage'
-export { default as HeadlineWithButton } from './HeadlineWithButton'
-export { default as InfoWithImage } from './InfoWithImage'
-export { default as TextInfo } from './TextInfo'
