+++ slices/EmailSignup/model.json
@@ -28,7 +28,12 @@
         },
         "inputLabel": {
           "type": "StructuredText",
-          "config": { "single": "paragraph", "label": "input_label" }
+          "config": {
+            "label": "Input Label",
+            "placeholder": "",
+            "allowTargetBlank": true,
+            "single": "paragraph"
+          }
         },
         "inputPlaceholder": {
           "type": "StructuredText",
@@ -39,9 +44,13 @@
         },
         "buttonText": {
           "type": "StructuredText",
-          "config": { "single": "paragraph", "label": "button_text" }
+          "config": {
+            "single": "paragraph",
+            "label": "button_text"
+          }
         }
-      }
+      },
+      "items": {}
     }
   ]
 }
