+++ slices/HeadlineWithButton/index.js
@@ -1,30 +1,22 @@
 import React from 'react'
-import { RichText } from 'prismic-reactjs'
+import { PrismicRichText } from '@prismicio/react'
 
 const HeadlineWithButton = ({ slice }) => (
-  <section>
-    <span className="title">
-      {
-        slice.primary.title ?
-        <RichText render={slice.primary.title}/>
-        : <h2>Template slice, update me!</h2>
-      }
-    </span>
-    {
-      slice.primary.description ?
-      <RichText render={slice.primary.description}/>
-      : <p>start by editing this slice from inside Prismic builder!</p>
-    }
-    <style jsx>{`
-        section {
-          max-width: 600px;
-          margin: 4em auto;
-          text-align: center;
-        }
-        .title {
-          color: #8592e0;
-        }
-    `}</style>
+  <section className="headline-with-button">
+    <div>
+      <PrismicRichText field={slice.primary.headline} />
+      <PrismicRichText field={slice.primary.description} />
+    </div>
+    <div className="button">
+      {slice.primary.button ? (
+        <img
+          src={slice.primary.button.url}
+          alt={slice.primary.button.alt || ''}
+        />
+      ) : (
+        <img alt="" />
+      )}
+    </div>
   </section>
 )
