+++ slices/InfoWithImage/index.js
@@ -1,31 +1,36 @@
 import React from 'react'
-import { RichText } from 'prismic-reactjs'
+import { PrismicRichText } from '@prismicio/react'
 
-const InfoWithImage = ({ slice }) => (
-  <section>
-    <span className="title">
-      {
-        slice.primary.title ?
-        <RichText render={slice.primary.title}/>
-        : <h2>Template slice, update me!</h2>
-      }
-    </span>
-    {
-      slice.primary.description ?
-      <RichText render={slice.primary.description}/>
-      : <p>start by editing this slice from inside Prismic builder!</p>
-    }
-    <style jsx>{`
-        section {
-          max-width: 600px;
-          margin: 4em auto;
-          text-align: center;
-        }
-        .title {
-          color: #8592e0;
-        }
-    `}</style>
-  </section>
-)
+const InfoWithImage = ({ slice }) => {
+  const featuredImage = slice.primary.featured_image
+  return (
+    <section className="info-with-image">
+      <div className="featured-image">
+        {featuredImage ? (
+          <picture>
+            <source
+              srcSet={featuredImage.mobile.url}
+              alt={featuredImage.mobile.alt || ''}
+              media="(max-width: 500px)"
+            />
+            <source
+              srcSet={featuredImage.tablet.url}
+              alt={featuredImage.tablet.alt || ''}
+              media="(max-width: 1100px)"
+            />
+            <img src={featuredImage.url} alt={featuredImage.alt || ''} />
+          </picture>
+        ) : (
+          <img alt="" />
+        )}
+      </div>
+      <div className="text-content">
+        <img src="/images/top-icon.png" alt="Checkbox icon" />
+        <PrismicRichText field={slice.primary.section_title} />
+        <PrismicRichText field={slice.primary.text} />
+      </div>
+    </section>
+  )
+}
 
 export default InfoWithImage
