+++ slices/TextInfo/index.js
@@ -1,30 +1,16 @@
 import React from 'react'
-import { RichText } from 'prismic-reactjs'
+import { PrismicRichText } from '@prismicio/react'
 
 const TextInfo = ({ slice }) => (
-  <section>
-    <span className="title">
-      {
-        slice.primary.title ?
-        <RichText render={slice.primary.title}/>
-        : <h2>Template slice, update me!</h2>
-      }
-    </span>
-    {
-      slice.primary.description ?
-      <RichText render={slice.primary.description}/>
-      : <p>start by editing this slice from inside Prismic builder!</p>
-    }
-    <style jsx>{`
-        section {
-          max-width: 600px;
-          margin: 4em auto;
-          text-align: center;
-        }
-        .title {
-          color: #8592e0;
-        }
-    `}</style>
+  <section className="text-info">
+    <div className="left-column">
+      <img src="/images/top-icon.png" alt="Checkbox icon" />
+      <PrismicRichText field={slice.primary.section_title} />
+      <PrismicRichText field={slice.primary.left_column_text} />
+    </div>
+    <div className="right-column">
+      <PrismicRichText field={slice.primary.right_column_text} />
+    </div>
   </section>
 )
