+++ slices/index.js
@@ -5,4 +5,3 @@ export { default as FullWidthImage } from './FullWidthImage'
 export { default as HeadlineWithButton } from './HeadlineWithButton'
 export { default as InfoWithImage } from './InfoWithImage'
 export { default as TextInfo } from './TextInfo'
-export { default as TextInfoTwo } from './TextInfoTwo'
